refactor: error 패키지 이름 exception 으로 변경
