+++ src/main/java/exception/AppException.java
@@ -1,4 +1,4 @@
-package error;
+package exception;
 
 public class AppException extends IllegalArgumentException {
     public AppException(final ErrorMessage errorMessage) {
